Create Exclamation_Analyzer3.py
This is a shorter piece of code which satisfies all given conditions. 
This was proposed to me upon review by a colleague. I show this code for other, more advanced users. 
My final piece of code (Analyzer2) is how I initially undertook the task, but this appears to satisfy the task just as well.